Fix zipline/basejump completed endpoint
This PR refactors and normalizes the endpoint and the ajax request.
Some weird bug keeps popping up on the server, but keymetrics has
decided to take a crap and not actually report back with the error
I don't know what exactly is causing it.

Normally I don't like fixing something I can't see, but I'm getting
a constant flood of emails because of some obscure bug here.

We removed the `completedWith` functionality a while back so that
was removed from this endpoint and the ajax call.

To test, verify that you can complete both a zipline/basejump.

Also show last updated date since it is possible a user might want
to update links