Fix Browser Zoom Related Issues
Some browsers have issues detecting exact font and image sizes when zoomed
in or out.  Prior workarounds have failed and we don't want to use regular
expressions, which have many defeciencies.

For the short term we will add a note to the affected challenges that the
browser zoom must be at 100%

Closes #6397
Closes #6242

Tested Locally