fix(travis-ci): Speed up the build for TravisCI
This commit speeds up the builds by upto 3 times than the current.
It uses npm 3 over npm 2 (travis default), and caches the
dependencies.

This is okay, because our prime two testing areas are, linting and
tests for the challenges.

NOTE: The speed up should happen from the second build,
as it will take atleast one build to cache the dependencies.