feat(app): Add Motivational Quote Easter Egg (#16932)
Added a feature where a motivational quote is randomly selected from a
static list and displayed whenever the user closes all the panes on the
page. Also added a quotes.js file with an array of quote objects and a
function that returns a random quote object when called.

Edit: Forgot to add double quotes to the quote blocks after moving from
test strings to random quote objects from the quotes.js file. Added
them here as HTML encoded left and right double quotes.

Made changes @Bouncey requested.

Made further changes @Bouncey requested. Also added a componentDidUpdate
method to both ensure the quote is only changed when panes open and close,
and also to make sure the new quote is not the same as the one that
was last displayed.

Removed the fCC logo as it was redundant and clashed with the night mode.
Simplified the layout so the quote block is in the middle of the page and
changed the width so that it's more readable. Also added styling to change
the quote text color in night mode.

BREAKING CHANGE: None

Closes #16382