Removed the base tag from the header of the preview iframe. The href and target attributes were causing links within the iframe to open a new tab and would point to /, which created problems with some challenges where users should link to other page elements like a footer.
Fixed an error found in QA where in-page anchors would work as expected in Chrome, but in FireFox would load another instance of Learn in the iframe. If the user goes back to the challenge, though, the in-page anchors work as expected. Readded <base> and had it point to href='' to get in-page anchors working in both Chrome and FF.