CSS: Add ::after pseudo-element document (#19381)
I noticed most of the pseudo-selectors/pseudo-elements are missing from this section, so I created a basic document as a starting point for the ::after pseudo-element. I tried to follow the same layout and style as the other articles in this section. This document is meant to be a start and added to/expounded on.

It's also my first real pull request - so please let me know how I can improve if needed 👍