Made several grammatical changes (#23288)
Removed space between "/" and "body" in line 27 to correct the spacing. 
Added "use" in between "container" and "several" to make grammatical sense in line 32. Also added a space between "semantic" and the parathesis in the same line to make text look cleaner.
Added "and" between "<header>" and "<footer>" because there are two items in line 33.

Note: In "Introduction of semantic tags," the grammar and sentence structure does not make sense. I tried to correct it, however, it may not be the right correction and should be reviewed.