Visual changes,  fix grammatical mistakes (#24952)
Split up Overview section, as it's rather long for a general overview. Removed "Unity is the engine used by Rust, Kerbal Space Program, and Cuphead" as it repeats pieces of information from the Noteworthy Games section. Also, removed the colon ":" from the Noteworthy Games title, as it's redundant with it being a header.