corrected typo (deleted the unnecessary backslash in the third example calculation) (#34292)
Change 1. corrected typo (deleted the unnecessary backslash in the third example calculation)
original version = -17.25\   ---> revised version = -17.25

Change 2. for greater clarity changed the word "great" to "useful" as follows:

The mean can be a great way to describe a set of numbers.  ---> The mean can be a useful way to describe a set of numbers.