List titles improved and bucle differences summary (#26905)
The translations seem literal and I have never seen them in this way. It is normal to speak directly of the reserved word that defines them.
I consider it important to say that there is a way to decide which of the four loops to choose and on what basis.