CodeSee take 2 (#42330)
* Adds CodeSee developer tooling to FreeCodeCamp client

CodeSee is a continuous code understanding platform,
meant to help developers when diving into a codebase.

This is an alpha version, which enables FCC contributors
to create and explore recordings of code execution on the
front-end. These recordings are meant to help orient a
developer to an unfamiliar part of the codebase, or help
diagnose when behavior is not matching our mental model.

When running in development, the client will now display
a small, purple CodeSee icon in the browser. Clicking this
icon starts (then stops) a new recording.

* chore(deps): update to @codesee/latest

* feat: switch to CodeSee hosted

* chore(deps): add process to root

I couldn't pin down what, exactly, but something in the client depends
on safer-buffer from the root, which tries to find process/browser, so
it needs installing here too.

Co-authored-by: Joshua Leven <josh@codesee.io>